fix: make vitest/browser types self-contained (fix #10280)
