+++ packages/vitest/browser/context.d.ts
@@ -1,7 +1,2 @@
-// @ts-ignore -- @vitest/browser-playwright might not be installed
-export * from '@vitest/browser-playwright/context'
-// @ts-ignore -- @vitest/browser-webdriverio might not be installed
-export * from '@vitest/browser-webdriverio/context'
-// @ts-ignore -- @vitest/browser-preview might not be installed
-export * from '@vitest/browser-preview/context'
-export { BrowserCommands, CDPSession, FsOptions } from 'vitest/internal/browser'
+export * from './context-base.js'
+export type { BrowserCommands, FsOptions } from 'vitest/internal/browser'
