+++ packages/vitest/browser/vendor-types.d.ts
@@ -1,0 +1,9 @@
+type AriaUtility = (...args: any[]) => any
+
+export const __ivyaAriaTypes: {
+  generateAriaTree: AriaUtility
+  renderAriaTree: AriaUtility
+  renderAriaTemplate: AriaUtility
+  parseAriaTemplate: AriaUtility
+  matchAriaTree: AriaUtility
+}
